# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sabor & Gestión API
  description: API REST del sistema Sabor & Gestión para la gestión operativa del restaurante.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
...
paths:
...
  /api/pagos/generar-qr/{pedidoId}:
    post:
      tags:
        - Pagos
      summary: Generar QR estático de pago
      description: Genera un código QR con ID del pedido y monto para cobro estático. Requiere autenticación.
      operationId: generarPagoQR
      security:
        - bearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      responses:
        '200':
          description: Código QR generado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerarPagoQRResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al generar QR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pagos/{pedidoId}/procesar:
    post:
      tags:
        - Pagos
      summary: Procesar pago final y cerrar orden
      description: 'Procesa el pago mediante una transacción atómica en MongoDB: cierra el pedido (CERRADO), crea el documento Pago, libera la mesa a "Libre" y genera el comprobante. Requiere rol Cajero o Administrador.'
      operationId: procesarPagoFinal
      security:
        - bearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcesarPagoRequest'
      responses:
        '200':
          description: Pago procesado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcesarPagoResponse'
        '400':
          description: Pedido ya cerrado o cancelado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Cajero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al procesar el pago
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pagos/notificar-qr/{pedidoId}:
    post:
      tags:
        - Pagos
      summary: Simular notificación de transferencia QR
      description: Endpoint público que simula la confirmación de transferencia QR desde un dispositivo móvil o webhook, emitiendo eventos WebSocket a la caja.
      operationId: simularPagoQR
      parameters:
        - name: pedidoId
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      responses:
        '200':
          description: Simulación exitosa y caja notificada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimularPagoQRResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al simular pago QR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
  /api/pagos/{pedidoId}/enviar-recibo:
    post:
      tags:
        - Pagos
      summary: Enviar recibo de pago por correo electrónico
      description: Genera el comprobante HTML detallado y lo envía al correo electrónico proporcionado mediante EmailService. Requiere rol Cajero o Administrador.
      operationId: enviarReciboCorreo
      security:
        - bearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          description: Identificador del pedido
          schema:
            type: string
          example: '{{pedidoId}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnviarReciboCorreoRequest'
      responses:
        '200':
          description: Recibo enviado por correo exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeRespuestaResponse'
              example:
                mensaje: Recibo enviado por correo exitosamente
        '400':
          description: Correo electrónico requerido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '403':
          description: Acceso denegado (requiere Cajero o Administrador)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '404':
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
        '500':
          description: Error al enviar el correo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MensajeErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Autenticación mediante token JWT en cabecera: Authorization: Bearer <token>'
  schemas:
    MensajeErrorResponse:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
    MensajeRespuestaResponse:
      type: object
      required:
        - mensaje
      properties:
        mensaje:
          type: string
          example: Operación realizada exitosamente
...
    GenerarPagoQRResponse:
      type: object
      properties:
        qrUrl:
          type: string
          example: https://api.qrserver.com/v1/create-qr-code/?size=250x250&data=SABOR_GESTION_ID_abc123_TOTAL_90
        total:
          type: number
          example: 90
    ProcesarPagoRequest:
      type: object
      properties:
        metodoPago:
          type: string
          enum:
            - Efectivo
            - Tarjeta
            - QR
          example: Efectivo
        porcentajeDescuento:
          type: number
          example: 0
        porcentajePropina:
          type: number
          example: 10
        montoDescuento:
          type: number
          example: 0
        montoPropina:
          type: number
          example: 9
        subtotalCierre:
          type: number
          example: 90
        cajeroAsignado:
          type: string
          example: '{{usuarioId}}'
    ComprobantePago:
      type: object
      properties:
        pedidoId:
          type: string
          example: '000000000000000000000070'
        meseroNombre:
          type: string
          example: Juan Pérez
        subtotal:
          type: number
          example: 90
        montoDescuento:
          type: number
          example: 0
        montoPropina:
          type: number
          example: 9
        descuentoAplicado:
          type: number
          example: 0
        propinaAplicada:
          type: number
          example: 9
        total:
          type: number
          example: 99
        totalPagado:
          type: number
          example: 99
        metodoPago:
          type: string
          example: Efectivo
        cajeroAsignado:
          type: string
          example: '000000000000000000000001'
        fechaBolivia:
          type: string
          example: 27/09/2026 15:00:00
        fecha:
          type: string
          format: date-time
    ProcesarPagoResponse:
      type: object
      properties:
        mensaje:
          type: string
          example: Pago procesado exitosamente
        comprobante:
          $ref: '#/components/schemas/ComprobantePago'
    SimularPagoQRResponse:
      type: object
      properties:
        exito:
          type: boolean
          example: true
        mensaje:
          type: string
          example: Simulación de pago exitosa. Notificando a la caja...
    EnviarReciboCorreoRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: consumidor@example.com
        clienteNombre:
          type: string
          example: Carlos Mendoza
        clienteCI:
          type: string
          example: '87654321'
...